% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_set.R
\name{as.hash_set}
\alias{as.hash_set}
\alias{as.hash_set.integer}
\alias{as.hash_set.character}
\title{Convert to a hash set}
\usage{
as.hash_set(x)

\method{as.hash_set}{integer}(x)

\method{as.hash_set}{character}(x)
}
\arguments{
\item{x}{An integer vector or a character vector.}
}
\value{
A hash set.
}
\description{
Convert to a hash set
}
\examples{
as.hash_set(1:10L)
as.hash_set(letters)
}
