% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_table.R
\name{as.hash_table}
\alias{as.hash_table}
\alias{as.hash_table.integer}
\alias{as.hash_table.numeric}
\alias{as.hash_table.character}
\alias{as.hash_table.logical}
\alias{as.hash_table.Date}
\alias{as.hash_table.POSIXct}
\alias{as.hash_table.list}
\title{Convert to a hash table}
\usage{
as.hash_table(x)

\method{as.hash_table}{integer}(x)

\method{as.hash_table}{numeric}(x)

\method{as.hash_table}{character}(x)

\method{as.hash_table}{logical}(x)

\method{as.hash_table}{Date}(x)

\method{as.hash_table}{POSIXct}(x)

\method{as.hash_table}{list}(x)
}
\arguments{
\item{x}{The input vector-like object with names.}
}
\value{
A hash table.
}
\description{
Convert to a hash table
}
\examples{
as.hash_table(c("a" = 1L, "b" = 2L))
as.hash_table(c("a" = 1, "b" = 2))
as.hash_table(c("a" = TRUE, "b" = FALSE))
as.hash_table(c("a" = "one", "b" = "two"))
as.hash_table(c("a" = as.Date("2025-01-01"), 
                "b" = as.Date("2025-02-01")))
as.hash_table(c("a" = as.POSIXct("2025-01-01 00:00:01"), 
                "b" = as.POSIXct("2025-02-01 00:00:01")))
as.hash_table(list("a" = 1:10, "b" = letters))
}
