% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generic.R
\name{generic}
\alias{generic}
\alias{hash_copy}
\alias{hash_delete}
\alias{hash_exists}
\alias{hash_insert}
\alias{hash_keys}
\alias{hash_size}
\alias{hash_values}
\title{Generic functions}
\description{
Generic functions
}
\details{
There are the following generic functions:
\itemize{
\item \code{hash_exists(h, keys)}: tests whether keys exist.
\item \code{hash_keys(h)}: returns a vector of keys.
\item \code{hash_values(h, keys = NULL)}: returns a vector of values.
\item \code{hash_size(h)}: returns the size of the hash table or hash set.
\item \code{hash_insert(h, keys, ...)}: inserts new keys or modify values for existing keys.
\item \code{hash_delete(h, keys)}: deletes keys.
\item \code{hash_copy(h)}: makes a copy of the hash table or hash set.
}

They can be used on the hash object created by \code{\link[=hash_table]{hash_table()}}, \code{\link[=hash_set]{hash_set()}}, \code{\link[=hash_fm_table]{hash_fm_table()}} and \code{\link[=hash_fm_set]{hash_fm_set()}}.
}
