% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_env.R
\name{hash_env_table}
\alias{hash_env_table}
\alias{hash_env_set}
\alias{hash_exists,hash_env-method}
\alias{hash_delete,hash_env-method}
\alias{hash_insert,hash_env_table-method}
\alias{hash_insert,hash_env_set-method}
\alias{hash_size,hash_env-method}
\alias{length.hash_env}
\alias{hash_keys,hash_env-method}
\alias{hash_values,hash_env_table-method}
\alias{hash_values,hash_env_set-method}
\alias{hash_copy,hash_env-method}
\alias{[[.hash_env}
\alias{[[<-.hash_env_table}
\alias{[.hash_env}
\alias{[<-.hash_env_table}
\alias{$.hash_env}
\alias{$<-.hash_env_table}
\alias{[[<-.hash_env_set}
\alias{[<-.hash_env_set}
\alias{$<-.hash_env_set}
\alias{as.hash_env}
\alias{as.hash_env.default}
\alias{as.vector.hash_env_set}
\alias{as.vector.hash_env_table}
\alias{as.list.hash_env_table}
\alias{show,hash_env_table-method}
\alias{show,hash_env_set-method}
\title{Hash table and hash set implemented by environment}
\usage{
hash_env_table(keys, values)

hash_env_set(keys)

\S4method{hash_exists}{hash_env}(h, keys)

\S4method{hash_delete}{hash_env}(h, keys)

\S4method{hash_insert}{hash_env_table}(h, keys, values)

\S4method{hash_insert}{hash_env_set}(h, keys)

\S4method{hash_size}{hash_env}(h)

\method{length}{hash_env}(x)

\S4method{hash_keys}{hash_env}(h)

\S4method{hash_values}{hash_env_table}(h, keys = NULL)

\S4method{hash_values}{hash_env_set}(h, keys = NULL)

\S4method{hash_copy}{hash_env}(h)

\method{[[}{hash_env}(x, i)

\method{[[}{hash_env_table}(x, i) <- value

\method{[}{hash_env}(x, i)

\method{[}{hash_env_table}(x, i) <- value

\method{$}{hash_env}(x, name)

\method{$}{hash_env_table}(x, i) <- value

\method{[[}{hash_env_set}(x, i) <- value

\method{[}{hash_env_set}(x, i) <- value

\method{$}{hash_env_set}(x, name) <- value

as.hash_env(x)

\method{as.hash_env}{default}(x)

\method{as.vector}{hash_env_set}(x, mode = "any")

\method{as.vector}{hash_env_table}(x, mode = "any")

\method{as.list}{hash_env_table}(x, ...)

\S4method{show}{hash_env_table}(object)

\S4method{show}{hash_env_set}(object)
}
\arguments{
\item{keys}{A character vector.}

\item{values}{An atomic vector or a list.}

\item{h, x, object}{A \code{hash_env_table} returned by \code{hash_env_table()} or a \code{hash_env_set} object returned by \code{hash_env_set()}.}

\item{i, name, value}{Keys and values.}

\item{mode, ...}{Please ignore.}
}
\description{
Hash table and hash set implemented by environment
}
\examples{
h = hash_env_table(letters, 1:26)
hash_keys(h)
hash_values(h)
h$a

h = hash_env_set(letters)
hash_exists(h, "a")
}
