% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_fm.R
\name{hash_fm_table}
\alias{hash_fm_table}
\alias{hash_fm_set}
\alias{hash_exists,hash_fm-method}
\alias{hash_delete,hash_fm-method}
\alias{hash_size,hash_fm-method}
\alias{length.hash_fm}
\alias{hash_keys,hash_fm-method}
\alias{hash_values,hash_fm_table-method}
\alias{hash_values,hash_fm_set-method}
\alias{hash_insert,hash_fm_table-method}
\alias{hash_insert,hash_fm_set-method}
\alias{hash_copy,hash_fm-method}
\alias{[[.hash_fm_table}
\alias{[[.hash_fm_set}
\alias{[[<-.hash_fm_table}
\alias{[.hash_fm_table}
\alias{[.hash_fm_set}
\alias{[<-.hash_fm_table}
\alias{$.hash_fm}
\alias{$<-.hash_fm_table}
\alias{[[<-.hash_fm_set}
\alias{[<-.hash_fm_set}
\alias{$<-.hash_fm_set}
\alias{as.hash_fm}
\alias{as.hash_fm.default}
\alias{as.vector.hash_fm_set}
\alias{as.vector.hash_fm_table}
\alias{as.list.hash_fm_table}
\alias{show,hash_fm_table-method}
\alias{show,hash_fm_set-method}
\title{Hash table and hash set implemented by fastmatch}
\usage{
hash_fm_table(keys, values)

hash_fm_set(keys)

\S4method{hash_exists}{hash_fm}(h, keys)

\S4method{hash_delete}{hash_fm}(h, keys)

\S4method{hash_size}{hash_fm}(h)

\method{length}{hash_fm}(x)

\S4method{hash_keys}{hash_fm}(h)

\S4method{hash_values}{hash_fm_table}(h, keys = NULL)

\S4method{hash_values}{hash_fm_set}(h, keys = NULL)

\S4method{hash_insert}{hash_fm_table}(h, keys, values)

\S4method{hash_insert}{hash_fm_set}(h, keys, values)

\S4method{hash_copy}{hash_fm}(h)

\method{[[}{hash_fm_table}(x, i)

\method{[[}{hash_fm_set}(x, i)

\method{[[}{hash_fm_table}(x, i) <- value

\method{[}{hash_fm_table}(x, i)

\method{[}{hash_fm_set}(x, i)

\method{[}{hash_fm_table}(x, i) <- value

\method{$}{hash_fm}(x, name)

\method{$}{hash_fm_table}(x, i) <- value

\method{[[}{hash_fm_set}(x, i) <- value

\method{[}{hash_fm_set}(x, i) <- value

\method{$}{hash_fm_set}(x, name) <- value

as.hash_fm(x)

\method{as.hash_fm}{default}(x)

\method{as.vector}{hash_fm_set}(x, mode = "any")

\method{as.vector}{hash_fm_table}(x, mode = "any")

\method{as.list}{hash_fm_table}(x, ...)

\S4method{show}{hash_fm_table}(object)

\S4method{show}{hash_fm_set}(object)
}
\arguments{
\item{keys}{A character vector. Keys should have no duplicates.}

\item{values}{An atomic vector or a list.}

\item{h, x, object}{A \code{hash_fm_table} object returned by \code{hash_fm_table()} or a \code{hash_fm_set} object returned by \code{hash_fm_set()}.}

\item{i, name, value}{Keys and values.}

\item{mode, ...}{Please ignore.}
}
\value{
\code{hash_fm_table()} returns a \code{hash_fm_table} object. \code{hash_fm_set()} returns a \code{hash_fm_set} object.
\code{hash_fm_delete()}, \code{hash_fm_insert()}, \code{hash_fm_copy()} return a \code{hash_fm_table} object
or a \code{hash_fm_set} object. \code{hash_exists()} returns a logical vector. \code{hash_size()} returns an integer. \code{hash_keys()} returns
a character vector. \code{hash_values()} on the \code{hash_fm_table} object returns a vector of a list which has the same format as in the
constructor function. \code{hash_values()} on the \code{hash_fm_set} object throws an error.
}
\description{
Hash table and hash set implemented by fastmatch
}
\details{
\code{hash_values()} and \code{[} methods on the \code{hash_fm_table} object preserve the original format of \code{values}, which means, if \code{values}
was specified as an atomic vector, the two functions also returns atomic vectors.

Once the hash table or the hash set is created, it is not allowed to modify.
}
\examples{
h = hash_fm_table(letters, 1:26)
hash_keys(h)
hash_values(h)
h$a
try(h$a <- 2L)

h = hash_fm_set(letters)
hash_exists(h, "a")
}
