% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_table.R
\name{hash_table}
\alias{hash_table}
\alias{hash_table_empty}
\title{Hash table}
\usage{
hash_table(keys, values)

hash_table_empty(cl)
}
\arguments{
\item{keys}{Keys as a character vector.}

\item{values}{Values. It supports vectors in the following classes: integer, logical, numeric, character, Date, POSIXct and list (i.e. a list of R objects).}

\item{cl}{One of "integer", "logical", "numeric", "character", "Date", "POSIXct" and "Robj".}
}
\value{
A \code{hash_table} object.
}
\description{
Hash table
}
\examples{
hash_table(c("a", "b"), 1:2L)
hash_table(c("a", "b"), c(TRUE, FALSE))
hash_table(c("a", "b"), c(0.1, 0.2))
hash_table(c("a", "b"), c("one", "two"))
hash_table(c("a", "b"), as.Date(c("2025-01-01", "2025-02-01")))
hash_table(c("a", "b"), as.POSIXct(c("2025-01-01 00:00:01", "2025-02-01 00:00:01")))
hash_table(c("a", "b"), list(1:10, letters))

# a nested hash table
h = hash_table_empty("Robj")
h$a = hash_table_empty("Robj")
h$a$b = hash_table("c", 1L)
h$a$b$c
hash_table_empty("integer")
hash_table_empty("character")
hash_table_empty("Robj")
}
