% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hash_table_keys}
\alias{hash_table_keys}
\alias{hash_table_values}
\alias{[[.hash_table}
\alias{[.hash_table}
\alias{$.hash_table}
\alias{[[<-.hash_table}
\alias{[<-.hash_table}
\alias{$<-.hash_table}
\alias{length.hash_table}
\alias{print.hash_table}
\title{Functions}
\usage{
hash_table_keys(h)

hash_table_values(h)

\method{[[}{hash_table}(x, i)

\method{[}{hash_table}(x, i)

\method{$}{hash_table}(x, name)

\method{[[}{hash_table}(x, i) <- value

\method{[}{hash_table}(x, i) <- value

\method{$}{hash_table}(x, name) <- value

\method{length}{hash_table}(x)

\method{print}{hash_table}(x, ...)
}
\arguments{
\item{h, x}{hash table}

\item{i, name}{keys}

\item{value}{Values}

\item{...}{Other arguments.}
}
\description{
Functions
}
