% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_set.R
\name{hash_set_keys}
\alias{hash_set_keys}
\alias{[[.hash_set}
\alias{[.hash_set}
\alias{$.hash_set}
\alias{[[<-.hash_set}
\alias{[<-.hash_set}
\alias{$<-.hash_set}
\alias{length.hash_set}
\alias{print.hash_set}
\title{High-level functions for hash sets}
\usage{
hash_set_keys(h)

\method{[[}{hash_set}(x, i)

\method{[}{hash_set}(x, i)

\method{$}{hash_set}(x, name)

\method{[[}{hash_set}(x, i) <- value

\method{[}{hash_set}(x, i) <- value

\method{$}{hash_set}(x, name) <- value

\method{length}{hash_set}(x)

\method{print}{hash_set}(x, ...)
}
\arguments{
\item{h, x}{A \code{hash_set} object.}

\item{i, name}{A single or a vector of keys.}

\item{value}{Values.}

\item{...}{Other arguments.}
}
\value{
\itemize{
\item \code{hash_set_keys()} returns a vector of all keys.
\item \code{[[} or \code{$} returns a single logical value.
\item \code{[} returns a vector of logical values.
}
}
\description{
High-level functions for hash sets
}
\details{
For the assignment, \code{value} should be logical. \code{TRUE} means to insert the corresponding keys to the hash set,
and \code{FALSE} means to delete the corresponding keys.
}
\examples{
h = hash_set(letters)
h$a
h[c("a", "b")]
h$a = FALSE
hash_set_keys(h)
h$aaaa = TRUE
hash_set_keys(h)
}
