% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_table.R
\name{hash_table_keys}
\alias{hash_table_keys}
\alias{hash_table_values}
\alias{[[.hash_table}
\alias{[.hash_table}
\alias{$.hash_table}
\alias{[[<-.hash_table}
\alias{[<-.hash_table}
\alias{$<-.hash_table}
\alias{length.hash_table}
\alias{print.hash_table}
\title{High-level functions for hash tables}
\usage{
hash_table_keys(h)

hash_table_values(h)

\method{[[}{hash_table}(x, i)

\method{[}{hash_table}(x, i)

\method{$}{hash_table}(x, name)

\method{[[}{hash_table}(x, i) <- value

\method{[}{hash_table}(x, i) <- value

\method{$}{hash_table}(x, name) <- value

\method{length}{hash_table}(x)

\method{print}{hash_table}(x, ...)
}
\arguments{
\item{h, x}{A \code{hash_table} object.}

\item{i, name}{A single or a vector of character keys.}

\item{value}{The corresponding values.}

\item{...}{Other arguments.}
}
\value{
\itemize{
\item \code{hash_table_keys()} returns a vector of all keys.
\item \code{hash_table_values()} returns a vector of all values.
\item \code{[[} or \code{$} returns a single value.
\item \code{[} returns a vector of values.
}
}
\description{
High-level functions for hash tables
}
\details{
For \code{[} and \verb{[<-} methods, multiple keys are allowed.

For the assignment, when the values in the hash table are not R object, if the assigned value is \code{NULL},
the corresponding key-value pair is deleted.
}
\examples{
h = hash_table(c("a", "b"), 1:2L)
hash_table_keys(h)
hash_table_values(h)
h[["a"]]
h$a
h[c("a", "b")]
h[["a"]] = 10L
h[["a"]]
h$a = 100L
h$a
length(h)
h = hash_table(c("a", "b"), 1:2L)
h$a = NULL
h
}
