% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hash_table_get_values}
\alias{hash_table_get_values}
\alias{hash_table_set_values}
\alias{hash_table_delete_pairs}
\alias{hash_table_size}
\alias{hash_table_get_all_keys}
\alias{hash_table_get_all_values}
\title{Low-level Functions}
\usage{
hash_table_get_values(h, keys)

hash_table_set_values(h, keys, values)

hash_table_delete_pairs(h, keys)

hash_table_size(h)

hash_table_get_all_keys(h)

hash_table_get_all_values(h)
}
\arguments{
\item{h}{hash table}

\item{keys}{keys}

\item{values}{values}
}
\description{
Low-level Functions
}
