% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_set.R
\name{hash_set_exists}
\alias{hash_set_exists}
\alias{hash_set_insert}
\alias{hash_set_delete}
\alias{hash_set_size}
\alias{hash_set_get_all_keys}
\alias{hash_set_clear}
\alias{hash_set_copy}
\title{Low-level functions for hash sets}
\usage{
hash_set_exists(h, keys)

hash_set_insert(h, keys)

hash_set_delete(h, keys)

hash_set_size(h)

hash_set_get_all_keys(h)

hash_set_clear(h)

hash_set_copy(h)
}
\arguments{
\item{h}{A \code{hash_set} object.}

\item{keys}{A single or a vector of keys.}
}
\value{
\itemize{
\item \code{hash_set_exists()} returns a logical vector.
\item \code{hash_set_insert()}/\code{hash_set_delete()}/\code{hash_set_clear()} directly modifies the input object \code{h}.
\item \code{hash_set_size()} returns a single integer.
\item \code{hash_set_get_all_keys()} returns a vector of all keys.
\item \code{hash_set_copy()} returns a new \code{hash_table} object.
}
}
\description{
Low-level functions for hash sets
}
\examples{
h = hash_set(letters)
hash_set_exists(h, "a")
hash_set_exists(h, "aa")
hash_set_insert(h, "aa")
hash_set_exists(h, "aa")
hash_set_delete(h, "aa")
}
