% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_table.R
\name{hash_table_get_values}
\alias{hash_table_get_values}
\alias{hash_table_exists}
\alias{hash_table_set_values}
\alias{hash_table_delete_pairs}
\alias{hash_table_size}
\alias{hash_table_get_all_keys}
\alias{hash_table_get_all_values}
\alias{hash_table_clear}
\alias{hash_table_copy}
\title{Low-level functions for hash tables}
\usage{
hash_table_get_values(h, keys)

hash_table_exists(h, keys)

hash_table_set_values(h, keys, values)

hash_table_delete_pairs(h, keys)

hash_table_size(h)

hash_table_get_all_keys(h)

hash_table_get_all_values(h)

hash_table_clear(h)

hash_table_copy(h)
}
\arguments{
\item{h}{A \code{hash_table} object.}

\item{keys}{A character vector of keys.}

\item{values}{The corresponding values.}
}
\value{
\itemize{
\item \code{hash_table_get_values()} returns a vector of values.
\item \code{hash_table_exists()} returns a logical vector.
\item \code{hash_table_set_values()}/\code{hash_table_delete_pairs()}/\code{hash_table_clear()} directly modifies the input object \code{h}.
\item \code{hash_table_size()} returns a single integer.
\item \code{hash_table_get_all_keys()} returns a vector of all keys.
\item \code{hash_table_get_all_values()} returns a vector of all values.
\item \code{hash_table_copy()} returns a new hash table.
}
}
\description{
Low-level functions for hash tables
}
\examples{
h = hash_table(c("a", "b"), 1:2L)
hash_table_get_values(h, "a")
hash_table_exists(h, c("a", "c"))
hash_table_set_values(h, "a", 100L)
hash_table_get_values(h, "a")
hash_table_delete_pairs(h, "a")
h
hash_table_clear(h)
h
}
