% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_set.R
\name{hash_set}
\alias{hash_set}
\alias{hash_set_empty}
\title{Hash set}
\usage{
hash_set(keys)

hash_set_empty(cl)
}
\arguments{
\item{keys}{An integer vector or a character vector where all elements are unique.}

\item{cl}{One of "integer" and "character".}
}
\value{
A \code{hash_set} object.
}
\description{
Hash set
}
\examples{
hash_set(letters)
hash_set(1:10)
hash_set_empty("integer")
hash_set_empty("character")
}
